requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
pymongo==4.6.1
dnspython==2.4.2
bcrypt==4.1.2
python-dotenv==1.0.0
gunicorn==21.2.0
